Support ECDH using OpenSSL-1.0.2
This is a hack to allow one to build libp11 and opensc's engine_pkcs11 using
OpenSSL-1.0.2. But OpenSSL-1.0.2 does not allow access to the ECDH_METHOD structure,
and doesn not provide access functions to change the compute_key function.
(OpenSSL-1.1 does provided the needed functions).

So this hack is only for use with 1.0.2 and only for those who are willing to debug
the code if OpenSSL changes anything that is hidden.

To gain access to the needed structure, the OpenSSL-1.0.2 src/crypto/ecdh/ech_locl.h
must be available while configuring libp11 and BUILD_ECDH_102 must be defined.

CPPFLAGS="-DBUILD_ECDH_102 -I/path/to/your/openssl-1.0.2/source/crypto/ecdh/"
export CPPFLAGS
configure ...

This patch may not complete, and has a least one memory leak.

The matching patch to engine_pkcs11 is also needed and requires:
CPPFLAGS="-DBUILD_ECDH_102"
export CPPFLAGS
configure...

It has been tested using the OpenSSL cms decrypt command.