Fix tests/pkcs11-uri-without-token.softhsm by adding id=
Added to tests/rsa-common.sh a pkcs11-tool test to show all slots.
In the pkcs11-uri-without-token.softhsm test there are two slots
one with objects and one uninitialized slot.

It appears that pkcs11-uri-without-token.softhsm may have been
trying to login to the uninitialized slot and failing as
some gdb testing showed.

So an id= was added to avoid this.

 On branch deffer_init_crypto3
 Changes to be committed:
	modified:   pkcs11-uri-without-token.softhsm
	modified:   rsa-common.sh